Clear the last pyright errors outside task_mdp
A dead domain_randomization kwarg that TypeErrors when joint-damping DR is
enabled (the fields now come from requires_model_fields), a randomizer whose
return value nothing reads and whose type the event decorator rejects, an
import of a TerrainImporterCfg mjlab no longer exports, bam actuator internals
the package does not declare, and get_base_metadata called through the wrapper
instead of the env.

pyright is now clean on src and tests.

+++ src/utils.py
@@ -71,7 +71,7 @@ def set_seed(seed: int = SEED, deterministic: bool = False) -> int:
     try:
         import warp as wp
 
-        wp.rand_init(seed)
+        wp.rand_init(wp.int32(seed))
     except Exception:
         pass
